requirements.txt
Flask==2.3.2
Flask-SocketIO==5.3.4
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-WTF==1.1.1
Werkzeug==2.3.7
eventlet==0.33.1
python-dotenv==1.0.0
